requirements.txt
PyJWT
Flask
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
python-dotenv
flask-jwt-extended
requests
